move code from game to score.js
Move code from game to score.js
